requirements.txt
# UI Framework
customtkinter==5.2.2
pillow>=10.3.0
tkinterdnd2>=0.3.0

# Video Processing
ffmpeg-python==0.2.0

# Transcription
openai-whisper>=20240930
google-genai>=1.0.0

# Audio Processing
pydub==0.25.1

# YouTube API
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
